Add missing instructions
I had to install dependencies before being able to run the first test. Without `npm install` the errors are different than what is stated in the excercise which can lead to confusion. 
As every excercise is it's own npm workspace, I assume that `npm install` needs to be run every time before testing a new excercise.